pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "fabric_ceph"
authors = [{name="Komal Thareja", email="kthare10@renci.org"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["Programming Language :: Python :: 3",
               "License :: OSI Approved :: MIT License",
               "Operating System :: OS Independent"]
description = "Fabric Ceph Service"
dynamic = ["version"]

keywords = ["Swagger", "Fabric  Ceph Service"]

requires-python = '>=3.9'
dependencies = [
    "requests >= 2.28.1",
    "waitress",
    "connexion==2.14.2",
    "swagger-ui-bundle>= 0.0.2",
    "PyYAML",
    "fabric_fss_utils==1.6.0",
    "fabric_fss_utils>=1.6.0",
    "fabric_credmgr_client>=1.6.2",
    "paramiko"
    ]

[project.optional-dependencies]
test = ["pytest",
        "flask_testing",
        "coverage>=4.0.3",
        "nose>=1.3.7",
        "pluggy>=0.3.1",
        "py>=1.4.31",
        "randomize>=0.13"
        ]

[project.urls]
Home = "https://fabric-testbed.net/"
Sources = "https://github.com/fabric-testbed/fabric_ceph"

[tool.flit.module]
name = "fabric_ceph"
